Update health variable data type to avoid issues if it goes negative

+++ src/Components/HealthComponent.hpp
@@ -2,6 +2,6 @@
 #include <cstdint>
 
 struct HealthComponent {
-  uint8_t health_amount;
-  HealthComponent(uint8_t health_amount = 0) : health_amount{health_amount} {}
+  int16_t health_amount;
+  HealthComponent(int16_t health_amount = 0) : health_amount{health_amount} {}
 };
